fix delete image gallerys

+++ src/gallery/gallery.service.ts
@@ -1,5 +1,5 @@
 // src/gallery/gallery.service.ts
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
@@ -16,9 +16,19 @@ export class GalleryService {
     });
   }
 
-  async deleteImage(id: number) {
+  async deleteImageFromPortfolio(portfolioId: number, imageId: number) {
+    const image = await this.prisma.galleryImage.findUnique({
+      where: { id: imageId },
+    });
+
+    if (!image || image.portfolioId !== portfolioId) {
+      throw new NotFoundException(
+        'تصویر موردنظر یافت نشد یا متعلق به این پروژه نیست',
+      );
+    }
+
     return this.prisma.galleryImage.delete({
-      where: { id },
+      where: { id: imageId },
     });
   }
